feat: Simplify and rename template functions
