+++ tools/openapi-gen/README.md
@@ -102,22 +102,20 @@ Templates have access to all [Sprig](https://masterminds.github.io/sprig/) funct
 | ---------------- | ----------------------------------- | ------------------------------------------------------ |
 | `schemaToGoType` | `schema → string`                   | Convert an OpenAPI schema to a Go type                 |
 | `paramToGoType`  | `param → string`                    | Convert an OpenAPI parameter to a Go type              |
-| `refToType`      | `ref → string`                      | Extract type name from a `$ref` string                 |
-| `isEnum`         | `schema → bool`                     | True if schema has enum values                         |
-| `isType`         | `schema, typeName → bool`           | True if schema matches the given OpenAPI type          |
-| `getType`        | `schema → string`                   | Return the OpenAPI type string                         |
-| `enumBaseType`   | `schema → string`                   | Underlying Go type for an enum (`string`, `int`, etc.) |
+| `refName`        | `ref → string`                      | Extract type name from a `$ref` string                 |
+| `getType`        | `schema → string`                   | Return the OpenAPI type string (nil-safe)              |
+| `enumBaseGoType` | `schema → string`                   | Underlying Go type for an enum (`string`, `int`, etc.) |
 | `enumValue`      | `schema, val → string`              | Format an enum constant value (quoted for strings)     |
 | `inlineEnums`    | `schemaName, schema → []inlineEnum` | Collect inline enum properties from a struct schema    |
 
 ### Property helpers
 
-| Function               | Signature                       | Description                                         |
-| ---------------------- | ------------------------------- | --------------------------------------------------- |
-| `propertyNamesOrdered` | `schemaName, schema → []string` | Property names in YAML source order                 |
-| `getProperty`          | `schema, name → *Schema`        | Get a property schema (traverses `allOf`)           |
-| `isRequired`           | `schema, name → bool`           | True if property is required (traverses `allOf`)    |
-| `isInOneOf`            | `schema, name → bool`           | True if property is defined inside a `oneOf` branch |
+| Function               | Signature                       | Description                                                    |
+| ---------------------- | ------------------------------- | -------------------------------------------------------------- |
+| `propertyNamesOrdered`  | `schemaName, schema → []string` | Property names in YAML source order                            |
+| `getProperty`           | `schema, name → *Schema`        | Get a property schema (traverses `allOf`)                      |
+| `getPropertyRequired`   | `schema, name → bool`           | True if property is required (traverses `allOf`)               |
+| `getTypePackage`        | `v → string`                    | Return `x-package` for a type ref (accepts `*Schema`, `*SchemaRef`, `*Parameter`, or `string`) |
 
 ### Iteration helpers
 
@@ -134,11 +132,8 @@ These return sorted slices for deterministic output:
 | Function         | Signature                      | Description                                    |
 | ---------------- | ------------------------------ | ---------------------------------------------- |
 | `capitalize`     | `string → string`              | Uppercase first letter                         |
-| `capitalizeEnum` | `string → string`              | Uppercase enum value for use in constant names |
-| `safeName`       | `string → string`              | Prefix Go reserved words with `_`              |
+| `goSafeName`     | `string → string`              | Prefix Go reserved words with `_`              |
 | `wrapComment`    | `text, width, prefix → string` | Word-wrap with prefix on continuation lines    |
-| `docsURL`        | `operation → string`           | Generate a documentation URL for an operation  |
-| `collectImports` | `schema → []string`            | Determine required Go imports for a schema     |
 
 ## Custom templates
