+++ tools/openapi-gen/templates/go-client/model.go.tmpl
@@ -15,13 +15,13 @@ var (
 {{ range .Models }}
 {{- $schema := .Schema }}
 {{- $schemaName := .SchemaName }}
-{{- if isEnum $schema }}
+{{- if $schema.Enum }}
 // {{ $schemaName }} defines the {{ $schemaName }} enum type.
-type {{ $schemaName }} {{ enumBaseType $schema }}
+type {{ $schemaName }} {{ enumBaseGoType $schema }}
 
 const (
 {{- range $schema.Enum }}
-	{{ $schemaName }}{{ capitalizeEnum (printf "%v" .) }} {{ $schemaName }} = {{ enumValue $schema . }}
+	{{ $schemaName }}{{ capitalize (replace "-" "_" (printf "%v" .)) }} {{ $schemaName }} = {{ enumValue $schema . }}
 {{- end }}
 )
 {{- else }}
@@ -35,7 +35,7 @@ type {{ $schemaName }} struct {
 {{- $prop := getProperty $schema . }}
 {{- if $prop }}
 {{- $propType := schemaToGoType $prop }}
-{{- $required := isRequired $schema . }}
+{{- $required := getPropertyRequired $schema . }}
 	{{ pascalcase . }} {{ if and (not $required) (not (hasPrefix "[]" $propType)) }}*{{ end }}{{ $propType }} `json:"{{ . }}{{ if not $required }},omitempty{{ end }}"`
 {{- end }}
 {{- end }}
@@ -43,11 +43,11 @@ type {{ $schemaName }} struct {
 {{- range $ie := inlineEnums $schemaName $schema }}
 
 // {{ $ie.Name }} defines the {{ $ie.Name }} enum type.
-type {{ $ie.Name }} {{ enumBaseType $ie.Schema }}
+type {{ $ie.Name }} {{ enumBaseGoType $ie.Schema }}
 
 const (
 {{- range $ie.Schema.Enum }}
-	{{ $ie.Name }}{{ capitalizeEnum (printf "%v" .) }} {{ $ie.Name }} = {{ enumValue $ie.Schema . }}
+	{{ $ie.Name }}{{ capitalize (replace "-" "_" (printf "%v" .)) }} {{ $ie.Name }} = {{ enumValue $ie.Schema . }}
 {{- end }}
 )
 {{- end }}
